Create Target
Most programmers will tell you that one of the ways to improve your performance in competitive programming is to practice a lot of problems. Ben took the above advice very seriously and decided to set a target for himself. Ben decides to solve at least 10 problems every week for 4 weeks and took an oath to do that. Given the number of problems he actually solved in each week over 4 weeks as P1, P2, P3, P4, determine if he fulfilled his oath or not.