Create Reach Home
Pavan is hungry and wants to reach home. Pavan can travel up to 5 kilometres on 1 litre of fuel on his motorcycle. Currently, his motorcycle is filled with X litres of fuel and his home is Y kilometres away. Determine whether Pavan can reach his home on his motorcycle or not.