Create Triangle Valid or Not
Write a Program that takes input as sides of a triangle and check whether the triangle is valid or not.
Hint:
A triangle is valid, if the sum of its two sides is greater than the third side. If three sides are a, b and c, then three conditions should be met.
1. a + b > c
2. a + c > b
3. b + c > a
Input Format
A single line contains three sides of a triangle a,b,c.