Create Day Name
Write a program to print days of week using if else ladder statement.
In this program, we will take a number between 1 to 7 as input from user, where 1 corresponds to Monday, 2 corresponds to Tuesday and so on. We will use if else ladder statement to print name of day in words.If any other input is given, then print "Invalid input".