Create Contest:
In a coding contest, there are two types of problems:
Easy problems, which are worth 1 point each
Hard problems, which are worth 2 points each
To qualify for the next round, a contestant must score at least X points. Professor solved A Easy problems and B Hard problems. Will Professor qualify or not?