Create Marathon
You are given four distinct integers a, b, c, d.
Bolt and three other people are running a Marathon. The value a is the distance that Bolt has run and b, c, d correspond to the distances the other three participants ran.
Output the number of participants in front of Bolt.