Create Culinary Eligibility
Kumar wants to organize a culinary competition for his students. However, he has made it mandatory that only students who have attended 75% or more of the culinary classes will be eligible to participate in the competition.
Given the total number of culinary classes conducted and the number of classes attended by a student, help Kumar to determine whether the student is eligible to participate in the culinary competition or not.