Create N - S - C
Take two numbers from user A and B respectively.
And print
n
sqaure of n (n2)
cube of n (n3)
where A < n < B
In other words print the NUMBER, SQUARE OF NUMBER and CUBE OF NUMBER for EVERY NUMBER from A to B (Both Exclusive). See the sample inputs for a better understanding.